+++ plugins/openshift-image-registry/dev/index.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { createDevApp } from '@backstage/dev-utils';
+import {
+  openshiftImageRegistryPlugin,
+  OpenshiftImageRegistryPage,
+} from '../src/plugin';
+
+createDevApp()
+  .registerPlugin(openshiftImageRegistryPlugin)
+  .addPage({
+    element: <OpenshiftImageRegistryPage />,
+    title: 'Root Page',
+    path: '/openshift-image-registry',
+  })
+  .render();