+++ plugins/openshift-image-registry/src/components/OcirImagesComponent/index.ts
@@ -1,0 +1,1 @@
+export { OcirImagesComponent } from './OcirImagesComponent';