+++ plugins/openshift-image-registry/src/plugin.test.ts
@@ -1,0 +1,7 @@
+import { openshiftImageRegistryPlugin } from './plugin';
+
+describe('openshift-image-registry', () => {
+  it('should export plugin', () => {
+    expect(openshiftImageRegistryPlugin).toBeDefined();
+  });
+});