+++ plugins/tekton/src/components/PipelineRunList/PipelineRunHeader.tsx
@@ -1,25 +1,36 @@
+import { sortable } from '@patternfly/react-table';
 import { tableColumnClasses } from './pipelinerun-table';
 
 const PipelineRunHeader = () => {
   return [
     {
-      title: 'Name',
+      title: 'NAME',
+      sortField: 'metadata.name',
+      transforms: [sortable],
       props: { className: tableColumnClasses[0] },
     },
     {
-      title: 'Status',
+      title: 'STATUS',
+      sortField: 'status.conditions[0].reason',
+      transforms: [sortable],
       props: { className: tableColumnClasses[1] },
     },
     {
-      title: 'Task status',
+      title: 'TASK STATUS',
+      sortField: 'status.conditions[0].reason',
+      transforms: [sortable],
       props: { className: tableColumnClasses[2] },
     },
     {
-      title: 'Started',
+      title: 'STARTED',
+      sortField: 'status.startTime',
+      transforms: [sortable],
       props: { className: tableColumnClasses[3] },
     },
     {
-      title: 'Duration',
+      title: 'DURATION',
+      sortField: 'status.completionTime',
+      transforms: [sortable],
       props: { className: tableColumnClasses[4] },
     },
   ];